+++ frontend/app/src/component/textArea/textArea.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 import "./textArea.css";
 
-export default function TextArea({label="label Propsに値を与えてください", value=""}) {
+export default function TextArea({label="label Propsに値を与えてください", defaultValue=""}) {
 
   return (
     <>
       <div className="input-area">
         <span className="input-label">{label}</span>
-        <textarea className="shadow" value={value} />
+        <textarea className="shadow" defaultValue={defaultValue} />
       </div>
     </>
   );
